嵌套MAP简单替换

+++ src/chipMaskHDF5.h
@@ -28,7 +28,7 @@ public:
     herr_t writeDataSet(std::string chipID, slideRange& sliderange, robin_hood::unordered_map<uint64, Position1>& bpMap, uint32_t BarcodeLen, uint8_t segment, uint32_t slidePitch, uint compressionLevel = 6, int index = 1);
     void openFile();
     void readDataSet(robin_hood::unordered_map<uint64, Position1>& bpMap, int index = 1);
-    void readDataSetSegment(robin_hood::unordered_map<uint64, bpmap_segment_value>& bpMapSegment, int index = 1);
+    void readDataSetSegment(robin_hood::unordered_map<uint32, bpmap_segment_value>& bpMapSegment, int index = 1);
 
 public:
     std::string fileName;
